requirements.txt
PyYAML==6.0.2
kafka-python>=2.0.2
prometheus_client>=0.17.1
